lp1-python/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.7